// src/app/shared/models/user/user-data.model.ts
export class UserData{
    id: string;
    userName: string;
    password: string;
    email: string;
    fullName: string;
    gender: number;
    avatar: string;
    dateOfBirth: Date;
    phoneNumber: string;
    address: string;
    lockoutEnabled: boolean;
    resetToken: string;
    resetTokenExpires: Date;
}